Fixed urls, and htmls. Added delete forms to all pages.

+++ orders/views.py
@@ -54,10 +54,15 @@ def order_update_detail(request, order_id, supplier_id, product_id):
     else:
         form = OrderDetailForm(instance=order_detail)
     return render(request, 'html/orderdetail_update.html', {'form': form, 'order': order_detail})
+
+
 def order_delete(request, order_id):
     order = get_object_or_404(Order, pk=order_id)
-    order.delete()
-    return redirect('order_list')
+    if request.method == 'POST':
+        order.delete()
+        return redirect('order_list')
+    return render(request, 'html/order_delete.html', {'order': order})
+
 def order_detail_delete(request, order_id, supplier_id, product_id):
     order_detail = get_object_or_404(OrderDetail, order_id=order_id, supplier_id=supplier_id, product_id=product_id)
     if request.method == 'POST':
